Refactored projectId page
